Add AddBidForm
